# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GST Invoice System API",
    "version": "1.0.0",
...
    "license": {
      "name": "See LICENSE in the repository"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://{host}",
      "description": "Your deployment",
      "variables": {
        "host": {
          "default": "invoices.example.com"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Set by POST /api/auth/login. HttpOnly, SameSite, and revocable server-side."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable message, safe to show a user."
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, present on errors a client is expected to handle (PLAN_LIMIT, PERIOD_LOCKED, VALUATION_RESTATEMENT, RETENTION_WARNING, DUPLICATE, HAS_MOVEMENTS, ALREADY_POSTED)."
          }
        }
      },
      "ValidationError": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "type": "object",
            "properties": {
              "issues": {
                "type": "array",
                "description": "Zod issues, when the failure was schema validation.",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ]
      },
      "Paise": {
        "type": "integer",
        "format": "int64",
        "description": "An amount in integer PAISE (1/100 rupee). Every monetary field in this API is paise and is named with a `Paise` suffix. Rupees are never sent over the wire: they cannot represent 1/3 of a rupee exactly, and accumulating them loses money.",
        "example": 118000
      },
      "Quantity": {
        "type": "number",
        "description": "A stock quantity. Fractional, because real units are (2.5 kg, 0.75 m). Not money - quantities are the only numbers in this API that are not integers.",
        "example": 2.5
      },
...
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "hsn": {
            "type": "string",
            "nullable": true
          },
          "barcode": {
            "type": "string",
            "nullable": true
          },
          "unit": {
            "type": "string",
            "example": "NOS"
          },
          "salePricePaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "purchasePricePaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "gstRate": {
            "type": "number",
            "example": 18
          },
          "cessRate": {
            "type": "number",
            "example": 0
          },
          "cessPerUnitPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "supplyType": {
            "type": "string",
            "enum": [
              "TAXABLE",
              "EXEMPT",
              "NIL_RATED",
              "NON_GST",
              "ZERO_RATED"
            ],
            "description": "These are four different things on a GST return even though three of them charge 0%. Do not collapse them."
          },
          "pricingMode": {
            "type": "string",
            "enum": [
              "EXCLUSIVE",
              "INCLUSIVE"
            ],
            "description": "INCLUSIVE means salePricePaise already contains GST (MRP goods)."
          },
          "openingStock": {
            "$ref": "#/components/schemas/Quantity"
          },
          "openingRatePaise": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Paise"
              }
            ],
            "description": "Cost of the opening stock. Separate from purchasePricePaise so that re-pricing an item does not retrospectively revalue stock already held."
          },
          "currentStock": {
            "$ref": "#/components/schemas/Quantity"
          },
          "lowStockAlert": {
            "$ref": "#/components/schemas/Quantity"
          },
          "trackBatches": {
            "type": "boolean"
          }
        }
      },
...
      "Ledger": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "groupId": {
            "type": "string"
          },
          "openingBalancePaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "openingIsDebit": {
            "type": "boolean"
          }
        }
      },
...
      "PaymentLink": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "token": {
            "type": "string",
            "description": "Unguessable public token, NOT the row id. The pay page is reachable without a session, so the identifier must not be enumerable."
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "EXPIRED",
              "CANCELLED"
            ]
          },
          "amountPaise": {
            "$ref": "#/components/schemas/Paise"
          },
          "paidPaise": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Paise"
              }
            ],
            "description": "Partial settlement leaves the link PENDING."
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/items": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "List stock and service items",
        "description": "Returns a plain array, not a paginated envelope.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Create an item",
        "description": "A non-zero `openingStock` also writes an OPENING stock movement, so the movement ledger stays the complete history of the item.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Item"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Read-only role, or the company is suspended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/payment-links": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "List payment links",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "1-based page number."
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "description": "Rows per page. Clamped to 200."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "page",
                    "pageSize",
                    "total",
                    "totalPages",
                    "hasMore"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PaymentLink"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create a payment link for an invoice",
        "description": "Reuses an open link for the same amount, and caps the amount at the outstanding balance.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "invoiceId"
                ],
                "properties": {
                  "invoiceId": {
                    "type": "string"
                  },
                  "amount": {
                    "oneOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "string"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLink"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Read-only role, or the company is suspended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ledgers": {
      "get": {
        "tags": [
          "Accounting"
        ],
        "summary": "Chart of accounts, grouped",
        "description": "Creates the standard chart on first access for companies that predate it.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Accounting"
        ],
        "summary": "Create a ledger",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ledger"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ledger"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Read-only role, or the company is suspended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}